Clean up inlinequery sending & make taghints easier to find
Simply type "#" to get a list of tags, and their help.
